+++ src/main/java/roomescape/time/controller/exception/ReservationTimeExceptionHandler.java
@@ -3,7 +3,7 @@ package roomescape.time.controller.exception;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.reservation.domain.exception.ReservationTimeNullException;
+import roomescape.time.domain.exception.ReservationTimeNullException;
 
 @RestControllerAdvice
 public class ReservationTimeExceptionHandler {
